+++ runtime/src/lib.rs
@@ -288,7 +288,7 @@ impl pallet_ds_accounts::Trait for Runtime {
 
 // After researches, consider placing here max grid sizes
 parameter_types! {
-    pub const MaxHeight: u16 = 400;
+    pub const MaxHeight: u32 = 400;
     pub const MaxBuildingsInArea: u16 = 100;
 }
